# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIKARIR API",
        "description": "Dokumentasi API untuk aplikasi Sikarir — Sistem Informasi Karir & Pelatihan BLK",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/jadwal-pelatihan": {
            "get": {
                "tags": [
                    "Jadwal Pelatihan"
                ],
                "summary": "Ambil semua jadwal pelatihan (paginate 15)",
                "operationId": "b438ccf7abf43622e1f83fb2bff22dbf",
                "responses": {
                    "200": {
                        "description": "Daftar jadwal pelatihan berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jadwal Pelatihan"
                ],
                "summary": "Buat jadwal pelatihan baru",
                "operationId": "d7eb78fdcdecd1c492c086ce30ec5ff3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pelatihan_id",
                                    "tanggal_mulai",
                                    "tanggal_selesai"
                                ],
                                "properties": {
                                    "pelatihan_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_mulai": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tanggal_selesai": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "lokasi": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "tersedia",
                                            "penuh",
                                            "selesai"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal pelatihan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jadwal-pelatihan/{id}": {
            "get": {
                "tags": [
                    "Jadwal Pelatihan"
                ],
                "summary": "Ambil detail jadwal pelatihan berdasarkan ID",
                "operationId": "a6fdccbf149ebbb1fb223746e08c5584",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail jadwal pelatihan"
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Jadwal Pelatihan"
                ],
                "summary": "Update jadwal pelatihan",
                "operationId": "aa2d317b9c1e89db6d533960f41a917a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tanggal_mulai": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tanggal_selesai": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "lokasi": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "tersedia",
                                            "penuh",
                                            "selesai"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil diupdate"
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jadwal Pelatihan"
                ],
                "summary": "Hapus jadwal pelatihan",
                "operationId": "d9116f3ace7b3563f4c3793363882e30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil dihapus"
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token yang diperoleh dari endpoint /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}